ViewRootImpl: Fixed memory leak caussed by viewrootImpl objects
When an application enters its own submenu and then switches to the background through
the HOME button, then viewrootImpl will increase when you enters the application again.
For example, settings module to perform the following operations:
settings (Application) ->Battery (submenu) ->HOME (virtual key) ->settings.
It is found that viewrootImpl increases and eventually leads to memory leak.
 We can observe the viewrootImpl object by shell command:dumpsys meminfo system

When the application is switched from the submenu to the background,
then you enters the application again,if there will be a failure to add view, which
leads to the reference viewrootImpl does not release.In fact,
the analysis of viewrootImpl objects from the MAT tool is mainly held by
mDisplayListener and mAccessibilityStateChangeListeners,
so we need to remove these objects when the view was added to failure.

Test: N/A

Change-Id: If995b4cfb6dedc8a8098cb787af9929b7b590972
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: ZeNiXxX <zenixxx.havoc@gmail.com>
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>
Signed-off-by: ben863 <r.budhiono@gmail.com>